modularize with QueryTypeSelector component
